requirements.txt
web3==6.15.0
python-telegram-bot==21.5